backend/requirements.txt
flask==3.1.1
flask-cors==5.0.1
supabase==2.13.0
python-dotenv==1.1.0
bcrypt==4.3.0
pyjwt==2.10.1
